moved epoch tracker query test
